pyproject.toml
[project]
name = "agentic-rag"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "faiss-cpu>=1.12.0",
    "fastapi>=0.115.0",
    "httpx>=0.28.1",
    "langchain>=0.3.27",
    "langchain-community>=0.3.27",
    "langchain-huggingface>=0.3.1",
    "openai>=1.99.9",
    "pytest>=8.4.1",
    "sentence-transformers>=5.1.0",
    "smolagents>=1.21.1",
    "uvicorn>=0.30.0",
]

[tool.uv]
dev-dependencies = [
    "pytest>=8.0.0",
    "httpx>=0.27.0",
]
